+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Added an exception with an error message if PyrEval is used with a single reference summary.
 
 ## [v0.2.0](https://github.com/danieldeutsch/sacrerouge/releases/tag/0.2.0) - 2021-03-26
 ### Added